// apps/ui/src/lib/api-client/models/SystemStatusResponse.ts
export interface SystemStatusResponse {
    /**
     * API version
     * @type {string}
     * @memberof SystemStatusResponse
     */
    version: string;
    /**
     * Timestamp (ISO 8601)
     * @type {string}
     * @memberof SystemStatusResponse
     */
    timestamp: string;
    /**
     * Server uptime in milliseconds
     * @type {number}
     * @memberof SystemStatusResponse
     */
    uptime: number;
    /**
     * Overall system status
     * @type {string}
     * @memberof SystemStatusResponse
     */
    overall: string;
    /**
     * @type {SystemStatusResponseFhir}
     * @memberof SystemStatusResponse
     */
    fhir: SystemStatusResponseFhir;
    /**
     * @type {SystemStatusResponseKeycloak}
     * @memberof SystemStatusResponse
     */
    keycloak: SystemStatusResponseKeycloak;
    /**
     * @type {SystemStatusResponseMemory}
     * @memberof SystemStatusResponse
     */
    memory: SystemStatusResponseMemory;
}